refactor: separate generation of shared object mapping into separate class
